// src/features/auth/auth.interface.ts
export interface UserInterface {
    id: string;
    name: string;
    email: string;
    username: string;
    password: string;
    createdAt: string;
    updatedAt: string;
}

export interface RegisterCredentialsInterface {
    id: string;
    name: string;
    surname: string;
    email: string;
    phoneNumber: string;
    username: string;
    password: string;
}